ipc: further fix service launcher
82a961b9d8dee3aa0d0a09e922fbb8588bd6febb had logical problems.
We'd never have arguments to supply from the user if we're calling this
function. And supplying them ourselves caused the service to expect
the service socket directory to already exist (which it usually won't).
Instead of telling the server what address to use and trusting it,
we use an address provided by the server and try to dial it.