parameters: remove redundant safety
reflect.Value.Set already does something similar internally.
We should be panicking in this case anyway, which is what reflect does.
It's an implementation fault and should never be allowed to happen at 
runtime.

We ditch the error context (which field this panic applies to), as this
is only useful when debugging the implementation. And remove the test 
for it, since we can't recover from a panic created inside goroutines.
(without complicating the implementation for no other reason)