archive: 9P "MOTD" file system experiment
Not actually good or useful. Just for evaluating the library.
Should be deleted later.

Exposes a read+write single level hierarchy, API, and CLI for simple
named text files.
Can be mounted, but IDs are hard coded for root.
Most things are implemented "just enough" to work, but are not spec
perfect. For example, no ".." support.