commands: partial revert of `printUsage` change
Keep the doc changes.
But revert change to interface.

The narrowing of the constraint here, is intentional.
This function currently does not call `.Write()`,
so io.StringWriter is enough to satisfy the call.

Currently only `FlagSet.SetOutput` needs the `io.Writer` portion.
Which is not within our control / stdlib.flag