PinFS: don't upgrade pins' CIDs from v0 -> v1
The original (pre-emptive) intention was to avoid (potential)
case-sensitive problems, stemming from a pin's CID.
While this is an okay idea, it's not our problem to solve, and doing so
incurs a cost per-request. This should be left up to the node operator.
If a CID v0 does cause problems, they can upgrade it (once), and use
the new v1 reference instead (removing the necessity of per-request
conversions).
This also removes the maintenance need to keep up to date with external
CID revisions (in this section of the code).