cgofuse: don't inline lookup tables / alloc once
Truthfully, I didn't benchmark this or look at the assembly.
But I remember this pattern being a concern in the past.

Unless the compiler has changed to optimize this out, things like the
map literal would be initialized each call.
Now we should do that only once on init.

Metrics would be nice to see on this though.
Does the compiler actually still act that way, and does the stack
climbing cost us more than doing that anyway?
Should these be lazy and initialized on the first call rather than at
init time?