9p: metadata - simplify + inline timestamp logic
This made more sense when every file system had its own SetAttr method.
Since we embed this type now, we can just inline the behavior.

We also drop the condition which returns an error when
`P9_SETATTR_XTIME_SET` is provided, but `P9_SETATTR_XTIME` is not.
Documentation doesn't specify what to do in this case, so instead, we
now silently drop the provided time value in such a request.
This check may be added back in later if it seems appropriate to do so.