command: reduce stack frames used for subcommands
The previous logic was correct, but wasteful.

Since Go does not do tail call optimization, each subcommand would incur
additional processing and stack frames, before executing the final
intended (sub)command.

Performance wise this is likely negligible. But when debugging a panic
this reduces stack frames for subcommands down to a single additional
frame.

This could likely be reduced even further with a trampoline, or
simple `become`|`goto func` if that ever gets added to Go.