9p: add `openFlags` type; extends `p9.OpenFlags`
The `p9.ReadOnly` constant is `0`. In order to distinguish between Go's
zero value for integer types, we can use an additional bit beyond the
`p9.OpenFlagsModeMask` to denote that a file was successfully
initialized (from within the `Open` method); and mask this bit out when
doing access checks (from within `Read`, `Write`, etc.).