9p: rework constructor helper implementations
This is more complicated than the old one, but should allow constructors
to be more flexible in how they construct complex data types.
It's also harder to misuse. If the source code falls out of sync, these
functions will panic with messages explaining what's wrong.
If the Go compiler had support for common struct access in unions, this
would be a compile time error `x.y = z` "x has no field y".
Since it does not, option functions should have Go tests to assure these
panics won't occur at runtime outside of development builds.

Option types can be easily shared among different type constructors by
using a struct type for the option's parameter.
The struct may contain fields (directly or embedded) that
will be assigned to when the option func is called.
E.g. `type settings struct { fieldName fieldType }`
Alternatively, the struct may contain a slice of option funcs, which is 
appended to when the option func is called. This allows a constructor
to accept options for a subtype's constructor.
E.g. `type settings { suboptions []optionType }`
Both styles can be used on the same struct at the same time, and values
will be propagated to all targets they apply to.
I.e. if an option can set a field on the struct itself, and it can be
appended to a slice on the struct; the value will be duplicated and
applied to both.