9p: tmp directory - call `Walk` before `UnlinkAt`
In 9P, `Tremove` can be considered a `Tclunk` with side effects.
In the 9P2000.L variant `Tunlinkat` is favoured in place of `Tremove`.
As such, we should have a reference derived from `Twalk` that we can
later `Tclunk` after requesting operations such as `Tunlinkat`.

Our `ephemeralDir` files depend on protocol compliance to function
correctly.
`ephemeralDir`s only unlink themselves from their parent when the last
reference to them is closed (after other conditions are also met).
Directly calling `UnlinkAt` circumvents the protocol by avoiding the
call to `Close` (which may be the final call to this reference).
This violation will cause a chain of ephemeral directories to remain
linked together when they are expected to unlink/recede towards to root.