9p: listener - fix invalid protocol for `UnlinkAt`
Walking prior to the `Tunlinkat` request assures our FID is valid and
`Tclunk` messages should (eventually) follow. In this case, after a
`Tunlinkat` request.

In the current implementation, calling `UnlinkAt` directly will succeed
for our local types, but this is a violation of the protocol that we do
not enforce internally. Remote/client files may not be as lax in their
implementation of the same interface, so we need to comply.
Internally this resolves an issue with `ephemeralDir` which expects the
protocol to be adhered to. Specifically in regard to `Close` being
called when the FID being operated on is finished.