commands: fix POSIX symbolic_mode parser
The previous version worked fine for "rwx" permissions, but did not
properly handle all the scenarios for `S_ISUID`, `S_ISGID`, `S_ISVTX`.
The current version passes some POSIX compliance tests and the code for
it is 100% covered, however, since some details of the specification are
implementation specific, the tests may fail when being tested against
`chmod` on platforms that have different behaviour and extension
differences.