9p: listener - restructure API
The listener system layout has changed.
Previously the structure was:
/root
  /$multiaddr-components[:-1]
    /$multiaddr-component[$last] <- represents the socket
i.e.
/listeners/
  /ip4
    /127.0.0.1
      /tcp
        /564 <- socket file

Now the layout is
/root
  /$multiaddr-components
    /listener <- socket file
    /connections
      /# <- incremental integer, represents an active connection
i.e.
/listeners/
  /ip4
    /127.0.0.1
      /tcp
        /564
	  listener
	  connections/
	    /1

Reading the contents of the `listener` file will return the multiaddr
for that listener.
Reading a connection file will return the local and remote addresses for
that connection, as well as some metrics. Currently represented in JSON.
E.g.
{
  "lastRead": "2023-05-11T13:42:09.385303-04:00",
  "lastWrite": "2023-05-11T13:42:09.385303-04:00",
  "local": "/ip4/192.168.1.40/tcp/564",
  "remote": "/ip4/192.168.1.40/tcp/52178",
  "#": 1
}

Both files contain the data necessary to reconstruct their own path.