9p: mount - change unmount API + refactor
Previously, each field of mount point file's data, was compared against
a list of "points" (provided as arguments to `UnmountTargets`).

Now, callers of `UnmountTargets` must also pass in a function which
parses the file's data and returns the "target" contained within.

Note that we don't define a particular data format for mount point
files, so coordination is left up to client implementations.

In our client, the `mount` CLI command receives arguments which are
encoded into the mount point file's data, which is written to the
server. `unmount` will call `UnmountTargets` expecting the data to be in
the same format as the data it send during `mount`. This should hold
true so long as no other client modified it between requests.

In addition, `UnmountAll` and `UnmountTargets` were refactored
internally, but their behaviour remains the same.