9p: mountpoint - change file format
In the Go API, we now require 2 methods in the `MountPoint` interface.
The methods return an identifier for a host and guest API, respectively.

In the 9P API, our mount point files were previously rendered as JSON
objects of this schema:
```
{
  "type": "object",
  "properties": {
    "host": {
      "type": "object"
    },
    "guest": {
      "type": "object"
    }
  }
}
```
The host and guest object schema are dependant on which host and guest
API is being used with this mount point.

Now, our mount point's encapsulate the previous object, under the name
"data", and add a "tag" object which contain identifiers for the host
and guest:
```
{
  "type": "object",
  "properties": {
    "data": {
      "type": "object",
      "properties": {
        "host": {
          "type": "object"
        },
        "guest": {
          "type": "object"
        }
      }
    },
    "tag": {
      "type": "object",
      "properties": {
        "host": {
          "type": "string"
        },
        "guest": {
          "type": "string"
        }
      },
      "required": [
        "host",
        "guest"
      ]
    }
  },
  "required": [
    "data",
    "tag"
  ]
}
```