commands: unmount - separate cmd flags from API options
The CLI command is only an entry point to the client's API call. It's
similar in name and values, but different in purpose. The CLI is only
responsible for parsing the command line argument vector, while the API
call does everything else. I.e. the command is responsible for parsing
strings and shuttling them to the API call if they were provided.

This change creates an hard separation between the 2.