9p: mount - address type concern regarding file creation
Previously, the mounts directory had a rigid format that had to be
conformed to. This is no longer the case. Our implementation still works
as a 3 level directory tree, but alternative versions could technically
be implemented with regular files, or whatever other scheme. Using our
specific directory types, their own types, or a combination.