commands: client - addr flag's default on new line
Depending on system and desktop-standards configuration (XDG, et al.),
this string can be long. Print it out on a line of its own.
Standard `flag` prepends a space, so we issue a backspace to remove it
when printing to a terminal.

This was already the case for the daemon command, but the client flags
were missing this in their definitions.