cgofuse: fix inconsistent key name
Go doesn't (currently) allow using constants in struct tags. Which makes
creates an explicit separation between these constants and the type
definitions they're associated with.
If the tags change again in the future, we should eat the cost
of using reflection retrieve the tags at runtime instead of hoping these
stay in sync in the source code.
Alternatively we could use a code generator, but the problem of making
sure it's run on change is essentially the same problem.