cgofuse: add mount point polling method
Instead of just waiting and assuming a mount succeeded, we insert a
special file into the `fs.FS` wrapper that we can query. If the file is
accessible, the mount should have succeeded and been fully initialized.

Previously, a timeout signaled success; since `Mount` only returns
quickly when it encounters an error, otherwise it blocks until `Unmount`
is called. Now, the timeout is treated as an error condition, implying
the mounted file system is not accessible by OS syscalls.

This should be more responsive than waiting on a fixed timer.