command: minor refine of `getSubcmmands`
This is a few instructions less, but is mainly just to deindent and give
more descript names to the innermost nested values.

In a previous version we needed to return `arguments` when recurring.
This was factored out at some point but the base cases were not updated.
We can also avoid the method call for some inputs.

An iterative version was considered, but seems more complex (in source
and asm). Was not benchmarked, but even if it turned out to be
marginally faster, it doesn't seem worth the change.
https://gist.github.com/djdv/3a6f206843fd841484d1de0c8c877823