commands: flag - allow symbolic name "nobody"
Printing the symbolic value "nobody" in the help text (while not
actually accepting that as a flag value) is misleading. However, we
don't want to print the actual ID value either. It's typically the max
value for the id type, which is large and potentially confusing also.

Instead, we can allow this symbolic "nobody" string value.
In the future, we might try to accept string values and do lookups for
system usernames & groups during flag parsing.