commands: adapt to `command` & `p9` API changes
This changes signatures of execute functions to accept a slice of
"options" instead of a type of "settings", effectively shifting the
responsibility of initializing default settings and applying
modifications; into the execute function itself, rather than relying on
the caller to do so correctly and consistently.
I.e. empty calls from the CLI and Go API should always have the same
behaviour.
In addition, lazy initialization can be done later (and more clearly),
if needed at all.

Help text has been populated and formatted in a way that should look
acceptable when printed on ASCII typewriters, as well as ANSI terminals.

Some of the constraints for `p9` changed slightly, so calls have been
updated.