ipfs: always get from cache (rather than peek)
Previously we were peeking the cache to avoid inflating the hit counter
in commonly seen sequences like `stat;stat;open`. However, it's not
likely to be detrimental / cause unwanted evictions if calls to `stat`
update the hit counter for a file. It may actually allow some paths to
fall out of cache if they're merely being queried/monitored frequently,
which is undesirable.