cgofuse: fix panic in `fileTable.Close` method
The table is expected to contain `nil` values as handles open and close
(to emulate the behaviour of Unix's lowest available `fd` semantics).
As a result, we need to detect and skip processing them during
`fileTable.Close`.
In particular, this was affecting `goWrapper.Destroy` which calls
`fileTable.Close`. The `cgofuse` library internally suppresses panics so
the effects of this were not visible when unmounting with `nil` handles.