commands: daemon - simplify stop sequence
Currently, the unmount phase waits for the server shutdown phase to
return. We do this to assure that connected clients don't instantiate
new mountpoints while the unmount phase is processing or after it has
already returned.
It's possible for these to be more independent, but since they depend on
sequential operation, we can drop a WaitGroup and a goroutine for now.