ipfs: all - fix ReadDir deadlock on EOF/error state
Originally, directory streams were struct values, and a channel field
was used to validate if the file was open / not-yet-closed.
Eventually, that was changed to be a pointer to struct, which was used
as a check condition instead. This invalidator path was never removed,
and causes deadlocks if the caller tries to read from a directory which
already encountered an error (either `EOF` or any other).