TODO:split + document; command: flag - refactor `flagSetFunc`
Previously, `flagSetFunc` would take in a function with parameters of
the value and container type,
which was responsible for inspecting and assigning the value.
Parsing of strings into values was handled internally by `flagSetFunc`.
This made it easier to combine our complex constructors with the [flag]
pkg. But forces us to handle parsing the string internally, among other
detriments.

Now, `flagSetFunc` was renamed to `flagOptionAppend`.
It takes in a string parser function, which produces a value, and a
function which receives a reference to the value's target memory
address.
This allows the caller to specify how the string argument should be
parsed, and where to store it on a successful parse.